+++ src/Controller/Manage/ProjectsController.php
@@ -22,6 +22,7 @@ use Eventum\Extension\ExtensionManager;
 use Eventum\Extension\Legacy\WorkflowLegacyExtension;
 use Eventum\Extension\RegisterExtension;
 use Eventum\Model\Entity;
+use Eventum\Model\Repository\ProjectRepository;
 use Eventum\ServiceContainer;
 use Project;
 use Status;
@@ -40,6 +41,8 @@ class ProjectsController extends ManageBaseController
 
     /** @var int */
     private $prj_id;
+    /** @var ProjectRepository */
+    private $repo;
 
     protected function configure(): void
     {
@@ -47,6 +50,7 @@ class ProjectsController extends ManageBaseController
 
         $this->cat = $request->request->get('cat') ?: $request->query->get('cat');
         $this->prj_id = $request->request->getInt('prj_id') ?: $request->query->getInt('prj_id');
+        $this->repo = Doctrine::getProjectRepository();
     }
 
     protected function defaultAction(): void
@@ -73,13 +77,12 @@ class ProjectsController extends ManageBaseController
             return;
         }
 
-        $repo = Doctrine::getProjectRepository();
         $project = $this->updateFromRequest(new Entity\Project(), $post);
         $project->setCreatedDate(new DateTime());
         $project->setAnonymousPost('disabled');
 
         try {
-            $repo->updateProject($project);
+            $this->repo->updateProject($project);
         } catch (DatabaseException $e) {
             $this->messages->addErrorMessage(ev_gettext('An error occurred while trying to add the new project.'));
 
@@ -123,10 +126,9 @@ class ProjectsController extends ManageBaseController
         }
 
         $prj_id = $post->getInt('id');
-        $repo = Doctrine::getProjectRepository();
-        $project = $this->updateFromRequest($repo->findOrCreate($prj_id), $post);
+        $project = $this->updateFromRequest($this->repo->findOrCreate($prj_id), $post);
         try {
-            $repo->updateProject($project);
+            $this->repo->updateProject($project);
         } catch (DatabaseException $e) {
             $this->messages->addErrorMessage(ev_gettext('An error occurred while trying to update the project information.'));
 
@@ -164,9 +166,7 @@ class ProjectsController extends ManageBaseController
     private function editAction(): void
     {
         $prj_id = $this->getRequest()->query->getInt('id');
-
-        $repo = Doctrine::getProjectRepository();
-        $project = $repo->findById($prj_id);
+        $project = $this->repo->findById($prj_id);
 
         $details = $project->toArray();
         $details['prj_assigned_users'] = Project::getUserColList($prj_id);